+++ xrpl_binary_codec/src/lib.rs
@@ -1,7 +1,6 @@
 //! Binary serialization for XRPL Protocol objects.
 
 mod error;
-pub mod field_id;
 pub mod parser;
 pub mod serialize;
 /// Implements serialization according to <https://xrpl.org/serialization.html>
