+++ xrpl_types/src/lib.rs
@@ -2,6 +2,7 @@
 //! WebSocket clients and potentially for server-side code.
 
 mod error;
+pub mod serialize;
 /// Types in internal canonical binary format <https://xrpl.org/serialization.html#type-list>
 mod types;
